Moves components into src directory
