+++ src/client/routes/dashboard.js
@@ -5,7 +5,7 @@ export  default {
   queries: ViewerQuery,
   getComponent(nextState, cb) {
     require.ensure([], (require) => {
-      cb(null, require('../components/Dashboard/DashboardContainer').default);
+      cb(null, require('../../components/Dashboard/DashboardContainer').default);
     });
   }
 };
